learn 74 page

+++ app/src/main/java/com/zx/kotlin_android/data/Coordinates.kt
@@ -1,3 +1,3 @@
 package com.zx.kotlin_android.data
 
-data class Coordinates(val lon: Long,val lat: Float)
+data class Coordinates(val lon: Float,val lat: Float)
